requirements.txt
langchain 
langchain_community 
langchain_groq 
python-dotenv
fastapi 
pydentic 
streamlit